Removed using business ID to find appointments.
This is a terrible solution that should never be done in the real-world. Unfortunately, we have a demo coming up and need this to work. For some reason, the API thinks it's a different business than it actually it. I suspect this has something to do with the /api/auth. Will investigate later.